Port to official docker client sdk

+++ test.sh
@@ -8,15 +8,33 @@ prefix=cdflow2-test-$RANDOM
 
 # run a registry - this is required because of https://github.com/moby/moby/issues/32016
 
-registry_id=$prefix-registry
+set +e
+registry_id=$(cat .test-registry-id 2>/dev/null)
+set -e
 
-docker run -d -p 5000 --name $registry_id registry:2 >/dev/null
+if [ -z "$PERSIST_REGISTRY" ]; then
+    if docker inspect "$registry_id" 2>/dev/null >/dev/null; then
+        docker stop $registry_id >/dev/null
+        docker rm $registry_id >/dev/null
+        rm .test-registry-id
+        registry_id=""
+    fi
+fi
 
-function finish {
-    docker stop $registry_id >/dev/null
-    docker rm $registry_id >/dev/null
-}
-trap finish EXIT 
+if [ -z "$registry_id" ]; then
+    registry_id=$prefix-registry
+    docker run -d -p 5000 --name $registry_id registry:2 >/dev/null
+    echo $registry_id > .test-registry-id
+fi
+
+if [ "$PERSIST_REGISTRY" == "" ]; then
+    function finish {
+        docker stop $registry_id >/dev/null
+        docker rm $registry_id >/dev/null
+        rm .test-registry-id
+    }
+    trap finish EXIT
+fi
 
 registry="localhost:$(docker inspect --format='{{(index (index .NetworkSettings.Ports "5000/tcp") 0).HostPort}}' "$registry_id")"
 
@@ -55,7 +73,7 @@ if [[ "$(go version)" != *"go1.13."* ]]; then
     exit 1
 fi
 
-tests="$(go list ./... | grep -v 'cdflow2$' | grep -v cdflow2/test)"
+tests="$(go list ./... | grep -v 'cdflow2$' | grep -v cdflow2/test | sort)"
 if [[ ! -z "$1" ]]; then
     tests="$(echo "$tests" | grep "$1")"
 fi
